Task-8: Refactor code
