+++ package.json
@@ -1,5 +1,6 @@
 {
   "type": "module",
+  "main": "server.js",
   "dependencies": {
     "axios": "^1.7.7",
     "date-fns": "^4.1.0",
